Implement Web3 Payment Gateway Integration in ASP.NET Core Backend #268

+++ src/Core/Data/Dto/ApplicationSettings/ApplicationSettingsDto.cs
@@ -4,11 +4,11 @@ namespace GamaEdtech.Data.Dto.ApplicationSettings
     {
         public int GridPageSize { get; set; } = 10;
         public string? DefaultTimeZoneId { get; set; }
-        public int SchoolContributionPoints { get; set; }
-        public int SchoolImageContributionPoints { get; set; }
-        public int SchoolCommentContributionPoints { get; set; }
-        public int PostContributionPoints { get; set; }
-        public int SchoolIssuesContributionPoints { get; set; }
-        public int RemoveSchoolImageContributionPoints { get; set; }
+        public long SchoolContributionPoints { get; set; }
+        public long SchoolImageContributionPoints { get; set; }
+        public long SchoolCommentContributionPoints { get; set; }
+        public long PostContributionPoints { get; set; }
+        public long SchoolIssuesContributionPoints { get; set; }
+        public long RemoveSchoolImageContributionPoints { get; set; }
     }
 }
