+++ src/Core/Data/Dto/Transaction/GetStatisticsResponseDto.cs
@@ -3,7 +3,7 @@ namespace GamaEdtech.Data.Dto.Transaction
     public sealed class GetStatisticsResponseDto
     {
         public string Name { get; set; }
-        public int DebitValue { get; set; }
-        public int CreditValue { get; set; }
+        public long DebitValue { get; set; }
+        public long CreditValue { get; set; }
     }
 }
